+++ bitcodin/test/encodingprofile/testcase_create_encoding_profile_cropping_config.py
@@ -17,7 +17,7 @@ class CreateEncodingProfileWithCroppingConfigTestCase(BitcodinTestCase):
     def runTest(self):
         audio_stream_config = AudioStreamConfig(default_stream_id=0, bitrate=192000)
         video_stream_config = VideoStreamConfig(default_stream_id=1, bitrate=512000,
-            profile='Main', preset='standard', height=480, width=640)
+            profile='Main', preset='premium', height=480, width=640)
 
         cropping_config = CroppingConfig(
             top=100,