+++ bitcodin/test/encodingprofile/testcase_create_encoding_profile_sample_rate.py
@@ -23,7 +23,7 @@ class CreateEncodingProfileWithSpecificSampleRatesTestCase(BitcodinTestCase):
             default_stream_id=1,
             bitrate=512000,
             profile='Main',
-            preset='standard',
+            preset='premium',
             height=480,
             width=640,
             rate=24