+++ bitcodin/test/job/testcase_create_job_deinterlace.py
@@ -29,7 +29,7 @@ class CreateJobWithDeinterlacing(BitcodinTestCase):
             default_stream_id=0,
             bitrate=512000,
             profile='Main',
-            preset='standard',
+            preset='premium',
             height=480,
             width=640
         )