+++ bitcodin/test/job/testcase_create_job_merge_audio_channels.py
@@ -24,7 +24,7 @@ class CreateJobTestCase(BitcodinTestCase):
                 default_stream_id=0,
                 bitrate=512000,
                 profile='Main',
-                preset='standard',
+                preset='premium',
                 height=480,
                 width=640
             )