+++ examples/create_job_encrypted_dash_and_hls.py
@@ -8,8 +8,31 @@ input_obj = bitcodin.Input(url='http://bitbucketireland.s3.amazonaws.com/Sintel-
 input_result = bitcodin.create_input(input_obj)
 
 video_configs = list()
-video_configs.append(bitcodin.VideoStreamConfig(default_stream_id=1, bitrate=512000, profile='Main',
-                                                preset='standard', height=480, width=640))
+
+video_configs.append(bitcodin.VideoStreamConfig(
+    default_stream_id=0,
+    bitrate=4800000,
+    profile='Main',
+    preset='premium',
+    height=1080,
+    width=1920
+))
+video_configs.append(bitcodin.VideoStreamConfig(
+    default_stream_id=0,
+    bitrate=2400000,
+    profile='Main',
+    preset='premium',
+    height=768,
+    width=1024
+))
+video_configs.append(bitcodin.VideoStreamConfig(
+    default_stream_id=0,
+    bitrate=1200000,
+    profile='Main',
+    preset='premium',
+    height=480,
+    width=854
+))
 
 audio_configs = [bitcodin.AudioStreamConfig(default_stream_id=0, bitrate=128000)]